EmployeeFactory added to create object
